backend/pyproject.toml
[project]
name = "agentpm"
version = "0.1.0"
description = "Local-first agent orchestration backend for feature planning and Claude Code execution"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.43",
  "pydantic>=2.11.7",
  "pydantic-settings>=2.11.0",
  "platformdirs>=4.4.0",
  "httpx>=0.28.1",
  "python-dotenv>=1.1.1",
  "pyyaml>=6.0.2",
  "mcp>=1.26.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "pyinstaller>=6.15.0",
]

[project.scripts]
agentpm-server = "agentpm.main:main"

[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["src"]
